Frontend/deepfake_detection/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
werkzeug==2.3.7
opencv-python==4.8.0.76
numpy==1.24.3
deepface==0.0.79
pyjwt==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0 